refactor(hermes): isolate agent control client
